fix(service): wire IterateObservations into all export format transforms
Moves IterateObservations out of the observations app layer into the
exports impl, since it's an internal helper for resolving observation
find specs (including field-filter search ids) rather than an
application-layer entry point requiring permission checks. Injects the
same instance into csv/kml/geojson/geopackage transforms from app.ts
so they share identical filter-resolution behavior instead of each
duplicating it.
