+++ service/src/app.impl/exports/app.impl.exports.csv.ts
@@ -3,7 +3,7 @@ import turfCentroid from '@turf/centroid'
 import * as json2csv from 'json2csv'
 import stream, { Readable } from 'stream'
 import { ExportParams, ExportTransform, LocationExportParams, ObservationExportParams } from '../../app.api/exports/app.api.exports'
-import { IterateObservations } from '../../app.api/observations/app.api.observations'
+import { IterateObservations } from './app.impl.exports'
 import { ExportItemSummary, ExportSummary } from '../../entities/exports/entities.exports'
 import { AttachmentStore, FormEntry, Observation, ObservationAttrs } from '../../entities/observations/entities.observations'
 import { UserLocation, UserLocationProperties, UserLocationRepository } from '../../entities/locations/entities.locations'
