+++ service/src/app.impl/exports/app.impl.exports.geojson.ts
@@ -2,7 +2,7 @@ import archiver, { Archiver } from 'archiver'
 import turfCentroid from '@turf/centroid'
 import stream, { Readable } from 'stream'
 import { ExportParams, ExportTransform, LocationExportParams, ObservationExportParams } from '../../app.api/exports/app.api.exports'
-import { IterateObservations } from '../../app.api/observations/app.api.observations'
+import { IterateObservations } from './app.impl.exports'
 import { ExportItemSummary, ExportSummary } from '../../entities/exports/entities.exports'
 import { Attachment, AttachmentStore, FormEntry, Observation, ObservationAttrs } from '../../entities/observations/entities.observations'
 import { UserLocationRepository } from '../../entities/locations/entities.locations'
