+++ service/src/app.impl/exports/app.impl.exports.geopackage.ts
@@ -5,7 +5,7 @@ import os from 'os'
 import moment from 'moment'
 import { Archiver } from 'archiver'
 import { ExportParams, ExportTransform, LocationExportParams, ObservationExportParams, ObservationFieldProjection } from '../../app.api/exports/app.api.exports'
-import { IterateObservations } from '../../app.api/observations/app.api.observations'
+import { IterateObservations } from './app.impl.exports'
 import { ExportItemSummary, ExportSummary } from '../../entities/exports/entities.exports'
 import { Attachment, AttachmentStore, Observation } from '../../entities/observations/entities.observations'
 import { UserLocationRepository } from '../../entities/locations/entities.locations'
