+++ service/src/app.impl/exports/app.impl.exports.kml.ts
@@ -1,5 +1,5 @@
 import { ExportParams, ExportTransform, LocationExportParams, ObservationExportParams } from '../../app.api/exports/app.api.exports'
-import { IterateObservations } from '../../app.api/observations/app.api.observations'
+import { IterateObservations } from './app.impl.exports'
 import { ExportItemSummary, ExportSummary } from '../../entities/exports/entities.exports'
 import { Attachment, AttachmentStore, FormEntry, Observation, ObservationAttrs } from '../../entities/observations/entities.observations'
 import { UserLocation, UserLocationRepository } from '../../entities/locations/entities.locations'
